+++ src/main/java/atdd/exception/SubwayExceptionHandler.java
@@ -1,4 +1,4 @@
-package atdd.station.exception;
+package atdd.exception;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
